+++ src/app/layout.tsx
@@ -1,5 +1,8 @@
 import { ExternalLink } from "lucide-react";
 import "@/lib/globals.css";
+import "@public/github-markdown-dark.css";
+import "@public/katex.min.css";
+import "@public/github-dark.css";
 import { inter } from "@public/fonts/fonts";
 import { BASE_URL, REPO_URL } from "@/lib/global";